+++ Model.py
@@ -1,84 +1,172 @@
-from ollama import Client
+from Settings import settings
+from Parameters import get_parameters
+from ollama import Client as Ollama_Client
+from openai import OpenAI as OpenAI_client
+import google.generativeai as gemini_client
+from llama_index.core import Settings, SimpleDirectoryReader
+from llama_index.core.base.llms.types import ChatResponse
+from llama_index.core.schema import ImageDocument
+from llama_index.llms.ollama import Ollama
+from llama_index.llms.openai import OpenAI
+from llama_index.llms.gemini import Gemini
+from llama_index.core.llms import ChatMessage
+from llama_index.multi_modal_llms.openai.utils import generate_openai_multi_modal_chat_message
+from llama_index.core.multi_modal_llms.generic_utils import image_documents_to_base64
 import wx
 from Utils import displayError
 from pathlib import Path
-import re
 import os
 from Parameters import get_parameters
 from RAG import RAG
-from Settings import settings
 import re
+import tiktoken
+import tiktoken_ext
+from tiktoken_ext import openai_public
+from llama_index.core.callbacks import CallbackManager, TokenCountingHandler
 
 class Model:	
-	def __init__(self, name="neural-chat", host="http://localhost:11434"):
-		self.host = host
-		self.client = Client(host=host)
+	def __init__(self):
 		self.messages = []
-		self.name = name
 		self.generate = False
 		self.image = None
 		self.rag = None
-		self.load_parameters()
+		self.models = []
+		self.token_counter = TokenCountingHandler(tokenizer=tiktoken.encoding_for_model("gpt-3.5-turbo").encode)
+		self.init_llm()
 
-	def updateRag(self):
-		self.rag.update_settings(self.host, self.name)
+	def get_models(self):
+		ids = []
+		if settings.llm_name == "Ollama":
+			ids = [model['name'] for model in Ollama_Client(host=settings.host).list()['models']]
+		if settings.llm_name == "OpenAI":
+			if not settings.openai_api_key: return ids
+			os.environ["OPENAI_API_KEY"] = settings.openai_api_key
+			client = OpenAI_client()
+			ids = [i.id for i in list(client.models.list().data) if i.id.startswith("gpt")]
+		if settings.llm_name == "Gemini":
+			if not settings.gemini_api_key: return ids
+			gemini_client.configure(api_key=settings.gemini_api_key)
+			ids = [m.name for m in list(gemini_client.list_models()) if 'generateContent' in m.supported_generation_methods]
+		ids.sort()
+		self.models = ids
+		return ids
+
+	def init_llm(self):
+		if settings.model_name not in self.models:
+			settings.model_name = self.get_models()[0]
+		print(settings.model_name)
+		options = get_parameters()
+		if settings.llm_name == "Ollama":
+			Settings.llm = Ollama(model=settings.model_name, request_timeout=600, base_url=settings.host, additional_kwargs=options)
+		if settings.llm_name == "OpenAI":
+			if not settings.openai_api_key: return
+			os.environ["OPENAI_API_KEY"] = settings.openai_api_key
+			additional_kwargs = {
+				"seed":options['seed'],
+				"temperature":options["temperature"],
+				"top_p":options["top_p"],
+				"max_tokens":options["num_ctx"],
+				"presence_penalty":options["presence_penalty"],
+				"frequency_penalty":options["frequency_penalty"],
+			}
+			Settings.llm = OpenAI(model = settings.model_name, additional_kwargs=additional_kwargs)
+		if settings.llm_name == "Gemini":
+			if not settings.gemini_api_key: return
+			os.environ["GOOGLE_API_KEY"] = settings.gemini_api_key
+			generate_kwargs = {
+				"temperature":options["temperature"],
+				"top_p":options["top_p"],
+				"top_k":options["top_k"],
+				"max_output_tokens":options["num_ctx"],
+			}
+			Settings.llm = Gemini(model_name=settings.model_name, generate_kwargs=generate_kwargs)
+
+		Settings.chunk_size = settings.chunk_size
+		Settings.chunk_overlap = settings.chunk_overlap
+		Settings.similarity_top_k = settings.similarity_top_k
+		Settings.similarity_cutoff = settings.similarity_cutoff
+		Settings.context_window = options['num_ctx']
+		#Settings.num_output = options['num_ctx']-256
+
+	def delete(self):
+		Ollama_Client(host=settings.host).delete(settings.model_name)
+
+	def create(self, name, modelfile):
+		Ollama_Client(host=settings.host).create(name, modelfile=modelfile, stream=False)
+
+	def modelfile(self):
+		return Ollama_Client(host=settings.host).show(settings.model_name)['modelfile']
+
+	def load_index(self, folder):
+		if not self.rag:
+			self.rag = RAG()
+		self.rag.load_index(folder)
 
 	def startRag(self, path, setStatus):
-		self.rag = RAG(self.host, self.name)
+		self.rag = RAG()
 		if isinstance(path, list): self.rag.loadFolder(path, setStatus)
 		elif path.startswith("http"): self.rag.loadUrl(path, setStatus)
 		else: self.rag.loadFolder(path, setStatus)
 		
-	def load_parameters(self):
-		self.parameters = get_parameters()
-		
 	def setHost(self, host):
-		self.host = host
-		self.client = Client(host=host)
-		if self.rag: self.updateRag()
+		settings.host = host
+		self.init_llm()
 
 	def setModel(self, name):
-		self.name = name
-		if self.rag: self.updateRag()
+		if settings.model_name == name: return
+		settings.model_name = name
+		self.init_llm()
 
 	def setSystem(self, system):
 		if system == "": return
-		system = {'role': 'system', 'content':system}
-		if len(self.messages) == 0 or self.messages[0]['role'] != "system":
+		system = ChatMessage(role='system', content=system)
+		if len(self.messages) == 0 or self.messages[0].role != "system":
 			self.messages.insert(0, system)
-		elif self.messages[0]['role'] == "system":
+		elif self.messages[0].role == "system":
 			self.messages[0] = system
 
 	def ask(self, content, window):
-		message = {'role': 'user', 'content': content}
+		self.token_counter.reset_counts()
+		if not self.image:
+			Settings.callback_manager = CallbackManager([self.token_counter])
+		message = ChatMessage(role='user', content=content)
 		if self.image:
-			message['images'] = [Path(self.image)]
-			self.image = None
-			for i in range(len(self.messages)):
-				if 'images' in self.messages[i]: self.messages[i].pop('images')
+			documents = SimpleDirectoryReader(input_files=[self.image]).load_data()
+			if settings.llm_name == "Ollama":
+				images = image_documents_to_base64(documents)
+				message = ChatMessage(role='user', content=content, additional_kwargs={'images':images})
+			elif settings.llm_name == "Gemini":
+				message = ChatMessage(role='user', content=content, additional_kwargs={'images':documents})
+			elif settings.llm_name == "OpenAI":
+				message = generate_openai_multi_modal_chat_message(prompt=content, role="user", image_documents=documents, image_detail="auto")
+			else:
+				print("Unknown")
 		try:
-			self.messages.append(message)
 			if content.startswith("/q ") and self.rag:
 				if not self.rag.index:
 					displayError(Exception("No index found."))
 					return
-				message['content'] = message['content'][3:]
+				message.content = message.content[3:]
 				self.messages.append(message)
 				wx.CallAfter(window.setStatus, "Processing with RAG...")
-				response = self.rag.ask(message['content'])
+				response = self.rag.ask(message.content)
 			else:
 				self.messages.append(message)
+				if settings.llm_name == "Gemini" and self.image: self.messages = self.messages[-1:]
 				wx.CallAfter(window.setStatus, "Processing...")
-				response = self.client.chat(model=self.name, messages=self.messages, stream=True, options=self.parameters)
-			message = ""
-			wx.CallAfter(window.response.AppendText, self.name[:self.name.index(":")].capitalize() + ": ")
+				response = Settings.llm.stream_chat(self.messages)
+			assistant_name = settings.model_name.capitalize()
+			if ":" in assistant_name:
+				assistant_name =  assistant_name[:assistant_name.index(":")]
+			wx.CallAfter(window.response.AppendText, assistant_name+": ")
 			self.generate = True
+			message = ""
 			sentence = ""
 			for chunk in response:
 				if not sentence: wx.CallAfter(window.setStatus, "Typing...")
 				data = chunk
 				if not isinstance(chunk, str):
-					chunk = chunk['message']['content']
+					chunk = chunk.delta
 				message += chunk
 				if settings.speakResponse:
 					sentence += chunk
@@ -97,8 +185,8 @@ class Model:
 					text = nodes[i].text
 					text = re.sub(r'\n+', '\n', text)
 					wx.CallAfter(window.response.AppendText, f"----------{os.linesep}Context {i+1} similarity score: {nodes[i].score:.2f}\n{text}{os.linesep}")
-
-			if 'total_duration' in data:
+			if isinstance(data, ChatResponse) and hasattr(data, 'raw') and 'total_duration' in data.raw:
+				data = data.raw
 				div = 1000000000
 				total = data['total_duration']/div
 				load = data['load_duration']/div
@@ -108,15 +196,19 @@ class Model:
 				gen_duration = data['eval_duration']/div
 				stat = f"Total: {total:.2f} secs, Load: {load:.2f} secs, Prompt: {prompt_count} tokens ({prompt_count/prompt_duration:.2f} t/s), Output: {gen_count} tokens ({gen_count/gen_duration:.2f} t/s)"
 				wx.CallAfter(window.setStatus, stat)
-			elif content.startswith("/q ") and self.rag:
-				message = f"Embedding Tokens: {self.rag.token_counter.total_embedding_token_count}, LLM Prompt Tokens: {self.rag.token_counter.prompt_llm_token_count}, LLM Completion Tokens: {self.rag.token_counter.completion_llm_token_count}, Total LLM Token Count {self.rag.token_counter.total_llm_token_count}"
-				wx.CallAfter(window.setStatus, message)
+			elif self.token_counter.total_llm_token_count:
+				status_message = f"Embedding Tokens: {self.token_counter.total_embedding_token_count}, LLM Prompt Tokens: {self.token_counter.prompt_llm_token_count}, LLM Completion Tokens: {self.token_counter.completion_llm_token_count}, Total LLM Token Count {self.token_counter.total_llm_token_count}"
+				wx.CallAfter(window.setStatus, status_message)
 			else:
-				wx.CallAfter(window.setStatus, "Finished.")
-			self.messages.append({"role":"assistant", "content":message.strip()})
+				wx.CallAfter(window.setStatus, "Finished")
+			if self.image:
+				self.messages[-1] = ChatMessage(role='user', content=content)
+			self.messages.append(ChatMessage(role="assistant", content=message.strip()))
 		except Exception as e:
 			self.messages.pop()
 			displayError(e)
 		finally:
 			self.generate = False
+			self.image = None
+			Settings.callback_manager = CallbackManager([])
 			wx.CallAfter(window.onStopGeneration)