+++ RAG.py
@@ -1,37 +1,18 @@
 from Settings import settings
 from llama_index.core import VectorStoreIndex, StorageContext, load_index_from_storage, SimpleDirectoryReader, Settings, get_response_synthesizer
-from llama_index.embeddings.huggingface import HuggingFaceEmbedding
+from llama_index.embeddings.ollama import OllamaEmbedding
 from llama_index.llms.ollama import Ollama
 from llama_index.readers.web import MainContentExtractorReader# TrafilaturaWebReader, BeautifulSoupWebReader, SimpleWebPageReader
 from llama_index.core.postprocessor import SimilarityPostprocessor
-from llama_index.core.callbacks import CallbackManager, TokenCountingHandler
-import tiktoken
 from Utils import displayError, displayInfo
 from time import time
-from tiktoken_ext import openai_public
-import tiktoken_ext
-from Parameters import get_parameters
 import os
 
 class RAG:
-	def __init__(self, host, model):
-		self.llm_name=model
-		self.host=host
-		#Settings.embed_model = OllamaEmbedding(base_url=host, model_name=model)
-		Settings.embed_model = HuggingFaceEmbedding(model_name=f"BAAI/bge-{settings.embed_model}-en-v1.5")
+	def __init__(self):
+		Settings.embed_model = OllamaEmbedding(base_url=settings.host, model_name="nomic-embed-text")
+		#Settings.embed_model = HuggingFaceEmbedding(model_name=f"BAAI/bge-{settings.embed_model}-en-v1.5")
 		self.index = None
-		self.token_counter = TokenCountingHandler(tokenizer=tiktoken.encoding_for_model("gpt-3.5-turbo").encode)
-		Settings.callback_manager = CallbackManager([self.token_counter])
-		self.update_settings(host, model)
-
-	def update_settings(self, host, model):
-		options = get_parameters()
-		Settings.llm = Ollama(model=model, request_timeout=600, base_url=host, additional_kwargs=options)
-		Settings.chunk_size = settings.chunk_size
-		Settings.chunk_overlap = settings.chunk_overlap
-		Settings.similarity_top_k = settings.similarity_top_k
-		Settings.similarity_cutoff = settings.similarity_cutoff
-		Settings.context_window = options['num_ctx']
 
 	def load_index(self, folder):
 		try:
@@ -73,8 +54,6 @@ class RAG:
 			setStatus("Failed to index.")
 
 	def ask(self, question):
-		self.token_counter.reset_counts()
-		self.update_settings(self.host, self.llm_name)
 		node_postprocessors = [SimilarityPostprocessor(similarity_cutoff=settings.similarity_cutoff)]
 		query_engine = self.index.as_query_engine(similarity_top_k=settings.similarity_top_k, node_postprocessors = node_postprocessors, response_mode='no_text')
 		response = query_engine.query(question)